+++ flowserv/model/database.py
@@ -10,6 +10,8 @@
 sessions as well as to create a fresh database.
 """
 
+import os
+
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, scoped_session
 
@@ -26,18 +28,18 @@ class DB(object):
     """Wrapper to establish a database connection and create the database
     schema.
     """
-    def __init__(self, connect_url=None, web_app=False, echo=False):
+    def __init__(self, connect_url=None, web_app=None, echo=False):
         """Initialize the database connection string. If no connect string is
         given an attempt is made to access the value in the respective
         environment variable. If the variable is not set an error is raised.
 
         Parameters
         ----------
-        connect_url: string, optional
+        connect_url: string, default=None
             Database connection string.
-        web_app: bool, optional
+        web_app: bool, default=None
             Use scoped sessions for web applications if set to True.
-        echo: bool, optional
+        echo: bool, default=False
             Flag that controlls whether the created engine is verbose or not.
 
         Raises
@@ -46,6 +48,14 @@ class DB(object):
         """
         # Ensure that the connection URL is set.
         connect_url = config.DB_CONNECT(value=connect_url)
+        # Get web_app flag from configuration.
+        web_app = config.WEBAPP(value=web_app)
+        # If the URL references a SQLite database ensure that the directory for
+        # the database file exists (Issue #68).
+        if connect_url.startswith('sqlite://'):
+            dbdir = os.path.dirname(connect_url[9:])
+            if dbdir:
+                os.makedirs(dbdir, exist_ok=True)
         if echo:
             import logging
             logging.info('Connect to database Url %s' % (connect_url))