+++ tests/app/test_run_app.py
@@ -92,7 +92,7 @@ def test_run_app_from_env(fsconfig, tmpdir):
     environment variables.
     """
     # -- Setup ----------------------------------------------------------------
-    os.environ[FLOWSERV_DB] = 'sqlite:///{}/flowserv.db'.format(str(tmpdir))
+    os.environ[FLOWSERV_DB] = 'sqlite:///{}/db/flowserv.db'.format(str(tmpdir))
     os.environ[FLOWSERV_API_BASEDIR] = str(tmpdir)
     os.environ[FLOWSERV_AUTH] = OPEN_ACCESS
     os.environ[FLOWSERV_ASYNC] = 'False'
@@ -100,8 +100,9 @@ def test_run_app_from_env(fsconfig, tmpdir):
     os.environ[FLOWSERV_FILESTORE_CLASS] = fsconfig[FLOWSERV_FILESTORE_CLASS]
     if FLOWSERV_S3BUCKET in os.environ:
         del os.environ[FLOWSERV_S3BUCKET]
-    from flowserv.app.database import flowdb
-    flowdb.init()
+    from flowserv.service.database import database
+    database.__init__()
+    database.init()
     app_key = install_app(source=TEMPLATE_DIR)
     # -- Run workflow ---------------------------------------------------------
     app = App(key=app_key)