+++ tests/app/test_test_env.py
@@ -15,8 +15,10 @@ from io import BytesIO, StringIO
 
 from flowserv.app.env import Flowserv
 from flowserv.controller.serial.docker import DockerWorkflowEngine
+from flowserv.model.files.base import FlaskFile
 from flowserv.model.files.fs import FSFile
 from flowserv.model.parameter.files import InputFile
+from flowserv.tests.files import FileStorage
 
 import flowserv.error as err
 import flowserv.model.workflow.state as st
@@ -72,6 +74,17 @@ def test_run_helloworld_in_env(source, specfile, filekey, tmpdir):
     assert not run.is_active()
     assert str(run) == st.STATE_SUCCESS
     assert len(run.files()) == 2
+    file_handles = dict()
+    for _, key, obj in run.files():
+        file_handles[key] = obj
+    assert file_handles[filekey]['name'] == filekey
+    assert file_handles[filekey]['title'] == 'Saying Hello to ...'
+    assert file_handles[filekey]['caption'] == 'Greetings for all persons.'
+    assert file_handles[filekey]['format'] == {'type': 'plaintext'}
+    assert file_handles['results/analytics.json']['name'] == 'results/analytics.json'  # noqa: E501
+    assert file_handles['results/analytics.json']['title'] == 'Score'
+    assert 'caption' not in file_handles['results/analytics.json']
+    assert file_handles['results/analytics.json']['format'] == {'type': 'json'}
     # -- Cancelling a finished run raises an error ----------------------------
     with pytest.raises(err.InvalidRunStateError):
         wf.cancel_run(run.run_id)
@@ -102,6 +115,7 @@ def test_run_helloworld_with_diff_inputs(tmpdir):
     files.append(filename)
     files.append(FSFile(filename))
     files.append(InputFile(FSFile(filename), 'output.txt'))
+    files.append(FlaskFile(FileStorage(FSFile(filename))))
     # -- Install and run the workflow -----------------------------------------
     wf = db.install(source=TEMPLATE_DIR, ignore_postproc=True)
     for file in files: