+++ tests/cli/conftest.py
@@ -17,7 +17,6 @@ from flowserv.cli.admin import cli
 from flowserv.config.api import FLOWSERV_API_BASEDIR
 from flowserv.config.database import FLOWSERV_DB
 from flowserv.config.backend import CLEAR_BACKEND
-from flowserv.model.base import WorkflowHandle
 from flowserv.service.api import service
 
 
@@ -30,17 +29,15 @@ def flowserv_cli(tmpdir):
     runner = CliRunner()
     os.environ[FLOWSERV_API_BASEDIR] = basedir
     os.environ[FLOWSERV_DB] = 'sqlite:///{}/flowserv.db'.format(basedir)
+    # Make sure to reset the database.
+    from flowserv.service.database import database
+    database.__init__()
     CLEAR_BACKEND()
     runner = CliRunner()
     runner.invoke(cli, ['init', '-f'])
     with service() as api:
         api.engine.fs = api.fs
     yield runner
-    # Using init does not seem to remove workflows from previous test runs.
-    # Need to delete them here explicitly:
-    from flowserv.service.database import database as db
-    with db.session() as session:
-        session.query(WorkflowHandle).delete()
     # Clear environment variables that were set for the test runner.
     del os.environ[FLOWSERV_API_BASEDIR]
     del os.environ[FLOWSERV_DB]